requirements.txt
Django==1.6
Unipath==1.0
distribute==0.6.34
dj-database-url==0.2.2
dj-static==0.0.5
django-decouple==1.1
gunicorn==18.0
static==0.4
wsgiref==0.1.2
psycopg2
python-social-auth
ipython
